requirements.txt
absl-py==0.8.0
alabaster==0.7.12
astor==0.8.0
atomicwrites==1.3.0
attrs==19.1.0
Babel==2.7.0
backcall==0.1.0
beautifulsoup4==4.8.1
bleach==1.5.0
bs4==0.0.1
certifi==2019.9.11
chainer==6.3.0
chainercv==0.13.1
chardet==3.0.4
Click==7.0
colorlog==4.0.2
coverage==4.5.3
cupy-cuda101==6.4.0
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
dlib==19.17.99
docutils==0.15.2
entrypoints==0.3
enum34==1.1.6
face-alignment==1.0.0
face-recognition==1.2.3
face-recognition-models==0.3.0
facenet-pytorch==0.3.1
fastrlock==0.4
filelock==3.0.12
gast==0.2.2
google-images-download==2.8.0
google-pasta==0.1.7
graphviz==0.13
grpcio==1.24.0
h5py==2.10.0
html5lib==0.9999999
idna==2.8
imageio==2.5.0
imageio-ffmpeg==0.3.0
imagesize==1.2.0
importlib-metadata==0.23
imutils==0.5.3
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jedi==0.15.1
Jinja2==2.10.1
joblib==0.14.0
json5==0.8.5
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-core==4.5.0
jupyterlab==1.1.1
jupyterlab-server==1.0.6
Keras==2.1.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keras-vggface==0.6
kiwisolver==1.1.0
Lasagne==0.1
llvmlite==0.30.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mistune==0.8.4
mock==3.0.5
more-itertools==7.2.0
moviepy==1.0.1
mtcnn==0.0.9
nbconvert==5.6.0
nbformat==4.4.0
networkx==2.3
notebook==6.0.1
numba==0.46.0
numpy~=1.18.1
numpydoc==0.9.2
opencv-python~=4.2.0.32
opt-einsum==3.1.0
packaging==19.2
pandas==0.24.2
pandocfilters==1.4.2
parso==0.5.1
pathlib2==2.3.5
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
pluggy==0.13.0
proglog==0.1.9
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.7.1
ptyprocess==0.6.0
py==1.8.0
pydot==1.4.1
Pygments==2.4.2
pyod==0.7.5
pyparsing==2.4.2
pyrsistent==0.15.4
pytest==5.2.0
python-dateutil==2.8.0
pytube==9.5.2
pytz==2019.2
PyWavelets==1.0.3
PyYAML==5.1.2
pyzmq==18.1.0
qtconsole==4.5.5
requests==2.22.0
scikit-image==0.15.0
scikit-learn~=0.22.1
scikit-neuralnetwork==0.7
scipy~=1.4.1
selenium==3.141.0
Send2Trash==1.5.0
six==1.12.0
snowballstemmer==2.0.0
soupsieve==1.9.4
Sphinx==2.3.1
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
tensorboard==1.14.0
tensorflow==1.4.0
tensorflow-estimator==1.14.0
tensorflow-tensorboard==0.4.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
Theano==1.0.4
torch~=1.4.0
torchvision==0.4.0
tornado==6.0.3
tqdm==4.36.1
traitlets==4.3.2
typing==3.6.6
typing-extensions==3.6.6
urllib3==1.25.6
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==0.6.0

cvt~=0.1.0
setuptools~=45.2.0
mmcv~=0.4.2